Move location of CookieBasedTuitionPartnerShortlistStorage
