+++ Application/GlobalUsings.cs
@@ -4,5 +4,3 @@ global using Application.Commands;
 global using Application.TuitionPartnerShortlistStorage.Interfaces;
 global using Domain.Search.ShortlistTuitionPartners;
 global using MediatR;
-global using Microsoft.AspNetCore.Http;
-global using Microsoft.Extensions.Logging;
