+++ UI/GlobalUsings.cs
@@ -1,0 +1,4 @@
+// Global using directives
+
+global using Application.TuitionPartnerShortlistStorage.Interfaces;
+global using Domain.Search.ShortlistTuitionPartners;
