+++ UI/Program.cs
@@ -1,8 +1,6 @@
 using System.Globalization;
 using System.Text.Json.Serialization;
 using Application.Extensions;
-using Application.TuitionPartnerShortlistStorage.Implementations;
-using Application.TuitionPartnerShortlistStorage.Interfaces;
 using FluentValidation;
 using FluentValidation.AspNetCore;
 using GovUk.Frontend.AspNetCore;
@@ -13,6 +11,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApplicationModels;
 using UI.Filters;
 using UI.Routing;
+using UI.Services.TuitionPartnerShortlistStorage;
 using AssemblyReference = UI.AssemblyReference;
 
 // Data import is a stand-alone process that should terminate once completed
